+++ aqua/core/fldstat/fldstat.py
@@ -1,11 +1,12 @@
 """AQUA class for field statitics"""
 import xarray as xr
 import numpy as np
+import regionmask
 
 from smmregrid import GridInspector
 
 from aqua.core.logger import log_configure, log_history
-from aqua.core.util import to_list, multiply_units
+from aqua.core.util import multiply_units
 
 from .area_selection import AreaSelection
 
@@ -17,10 +18,10 @@ class FldStat():
     """AQUA class for field statitics"""
 
     def __init__(self,
-             area: xr.Dataset | xr.DataArray | None = None,
-             horizontal_dims: list[str] | None = None,
-             grid_name: str | None = None,
-             loglevel: str = 'WARNING'):
+                 area: xr.Dataset | xr.DataArray | None = None,
+                 horizontal_dims: list[str] | None = None,
+                 grid_name: str | None = None,
+                 loglevel: str = 'WARNING'):
         """
         Initialize the FldStat.
 
@@ -46,7 +47,7 @@ class FldStat():
             raise ValueError("Area must be an xarray DataArray or Dataset.")
 
         self.grid_name = grid_name
-        
+
         # Initialize area selection
         self.area_selection = AreaSelection(loglevel=loglevel)
 
@@ -58,6 +59,9 @@ class FldStat():
 
     def fldstat(self, data: xr.DataArray | xr.Dataset,
                 stat: str = "mean",
+                region: regionmask.Regions | None = None,
+                region_sel: str | int | list | None = None,
+                mask_kwargs: dict = {},
                 lon_limits: list | None = None, lat_limits: list | None = None,
                 dims: list | None = None,
                 **kwargs):
@@ -68,6 +72,9 @@ class FldStat():
         Args:
             data (xr.DataArray or xarray.DataDataset):  the input data
             stat (str):  the statistic to compute, only supported is "mean"
+            region (regionmask.Regions, optional): A regionmask Regions object defining a class regions.
+            region_sel (str, int or list, optional): The region(s) to select by name or number from the region object.
+            mask_kwargs (dict, optional): Additional keyword arguments passed to region.mask().
             lon_limits (list, optional):  the longitude limits of the subset
             lat_limits (list, optional):  the latitude limits of the subset
             dims (list, optional):  the dimensions to average over, if not provided, horizontal_dims are used
@@ -107,8 +114,8 @@ class FldStat():
             for dim in dims:
                 if dim not in self.horizontal_dims:
                     raise ValueError(f"Dimension {dim} not found in horizontal dimensions: {self.horizontal_dims}")
-        
-        #if area is not provided, return the raw mean
+
+        # If area is not provided, return the raw mean
         if self.area is None:
             self.logger.warning("No area provided, no area-weighted stat can be provided.")
             # compact call, equivalent of "out = data.mean()"
@@ -122,9 +129,12 @@ class FldStat():
 
         # align coordinates values of area to match data
         self.area = self.align_area_coordinates(data)
-
-        if lon_limits is not None or lat_limits is not None:
+    
+        if lon_limits is not None or lat_limits is not None or region is not None:
+            self.logger.debug("Selecting area for field stat calculation.")
             data = self.area_selection.select_area(data, lon=lon_limits, lat=lat_limits,
+                                                   region=region, region_sel=region_sel,
+                                                   mask_kwargs=mask_kwargs,
                                                    to_180=False, **kwargs)
 
         # cleaning coordinates which have "multiple" coordinates in their own definition
@@ -144,16 +154,19 @@ class FldStat():
         else:
             weighted_data = data.weighted(weights=self.area.fillna(0))
             out = getattr(weighted_data, stat)(dim=dims)
-        
+
         if self.grid_name is not None:
             log_history(out, f"From grid '{self.grid_name}'. Computed field stat '{stat}'")
 
         return out
-    
+
     def select_area(self, data: xr.Dataset | xr.DataArray,
                     lon: list | None = None, lat: list | None = None,
                     box_brd: bool = True, drop: bool = False,
                     lat_name: str = "lat", lon_name: str = "lon",
+                    region: regionmask.Regions | None = None,
+                    region_sel: str | int | list | None = None,
+                    mask_kwargs: dict = {},
                     default_coords: dict = {"lat_min": -90, "lat_max": 90,
                                             "lon_min": 0, "lon_max": 360},
                     to_180: bool = True) -> xr.Dataset | xr.DataArray:
@@ -165,12 +178,14 @@ class FldStat():
         # select_area method arguments. However it is possible to foresee
         # that we may want to automatically detect the names
         return self.area_selection.select_area(data, lon=lon, lat=lat,
-                                                box_brd=box_brd, drop=drop,
-                                                lat_name=lat_name, lon_name=lon_name,
-                                                default_coords=default_coords, to_180=to_180)
+                                               box_brd=box_brd, drop=drop,
+                                               lat_name=lat_name, lon_name=lon_name,
+                                               region=region, region_sel=region_sel,
+                                               mask_kwargs=mask_kwargs,
+                                               default_coords=default_coords, to_180=to_180)
 
-    def integrate_over_area(self, data: xr.Dataset | xr.DataArray, 
-                            areacell: xr.DataArray, 
+    def integrate_over_area(self, data: xr.Dataset | xr.DataArray,
+                            areacell: xr.DataArray,
                             dims: list):
         """
         Compute the integral of the data over the area.
@@ -179,12 +194,12 @@ class FldStat():
             data (xr.DataArray or xr.Dataset): The data, used also for masking.
             areacell (xr.DataArray): The area cells.
             dims (list): Dimensions to sum over.
-            
+
         Returns:
             xr.DataArray or xr.Dataset: The integral of the data over the area
         """
         area_weighted_data = data * areacell.where(data.notnull())
-    
+
         # preserve attrs (e.g. AQUA_region) from areacell due to multiplication above, which has priority if keys overlap
         merged_attrs = {**data.attrs, **areacell.attrs}
 
@@ -193,39 +208,39 @@ class FldStat():
         else:
             self.logger.warning(f"Data units: {data.attrs.get('units', 'None')}; "
                                 f"Area units: {areacell.attrs.get('units', 'None')}, cannot multiply units using Metpy.")
-        
+
         if 'long_name' in data.attrs:
             merged_attrs['long_name'] = f"Integrated {data.attrs['long_name']}"
-        
+
         area_weighted_data.attrs.update(merged_attrs)
 
         area_weighted_integral = area_weighted_data.sum(skipna=True, min_count=1, dim=dims)
-        
+
         return area_weighted_integral
 
-    def sum_area(self, 
+    def sum_area(self,
                  data: xr.Dataset | xr.DataArray, 
                  areacell: xr.DataArray, 
                  dims: list):
         """
         Compute the sum of area cells where masked data is not null.
-        
+
         This is useful for computing field such as sea ice extent, by summing
         the area of cells that contain data not null.
 
         Note: if data is not masked might return incorrect results. If irrelevant regions (e.g., low level or land
         in sea-ice data) are not masked beforehand, their area will be incorrectly included in the sum.
-        
+
         Args:
             data (xr.DataArray or xr.Dataset): The data (check if pre-masking is needed in the considered variable)
             areacell (xr.DataArray): The area cells.
             dims (list): Dimensions to sum over.
-            
+
         Returns:
             xr.DataArray or xr.Dataset: The sum of area cells
         """
         summed_area = areacell.where(data).sum(skipna=True, min_count=1, dim=dims)
-        
+
         if self.grid_name is not None:
             log_history(summed_area, f"Area summed from {self.grid_name} grid")
         return summed_area
@@ -301,9 +316,9 @@ class FldStat():
                 if np.array_equal(area_rounded, data_rounded):
                     self.logger.warning("Coordinate '%s' aligned by rounding to %d decimals.", coord, decimals)
                     # assign the rounded coordinates to the area (matching data's rounded values)
-                    self.area = self.area.assign_coords({coord: data_rounded})
+                    self.area = self.area.assign_coords({coord: data_coord})
                     continue
 
                 raise ValueError(f"Mismatch in values for coordinate '{coord}' between data and areas.")
-    
+
         return self.area