+++ CHANGELOG.md
@@ -5,9 +5,26 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 ## [Unreleased]
 
-Unreleased in the current development version (target v0.22.0):
+Unreleased in the current development version (target v1.0.0):
+
+Complete list:
+- Explicit netcdf4 dependency in conda-forge and dask-distributed in pyproject to reduce incompatibilities (#2603)
+- Remove support for intake-esm due to deprecation of intake v1 version (#2603)
+- Aqua web push folder management improvements (#2574)
+- Improve fldstat handling of gaussian grids (#2584)
+- Data model now uses a ranking system to identify coordinates (#2585) 
+
+## [v1.0.0a1]
+
+Main changes:
+1. New `show_catalog_content` function deprecating `catalog` and `inspect_catalog`
+2. Shapefile support for spatial selection in `AreaSelection`
+3. Documentation aligned to the new repository structure
 
 AQUA core complete list:
+- Allow to show descriptions in `show_catalog_content` (#2589)
+- Improve test cleaning and avoid specific race conditions in tests (#2587)
+- Shapefile support for spatial selection in `AreaSelection` (#2576)
 - Remove IPython dependency from io_util (#2569)
 - Add LaTeX units formatting to labels (#2561)
 - Add concurrency suppression for Github actions (#2557)
@@ -1309,7 +1326,8 @@ This is mostly built on the `AQUA` `Reader` class which support for climate mode
 This is the AQUA pre-release to be sent to internal reviewers. 
 Documentations is completed and notebooks are working.
 
-[unreleased]: https://github.com/DestinE-Climate-DT/AQUA/compare/v0.21.0...HEAD
+[unreleased]: https://github.com/DestinE-Climate-DT/AQUA/compare/v1.0.0a1...HEAD
+[v1.0.0a1]: https://github.com/DestinE-Climate-DT/AQUA/compare/v0.21.0...v1.0.0a1
 [v0.21.0]: https://github.com/DestinE-Climate-DT/AQUA/compare/v0.20.0...v0.21.0
 [v0.20.0]: https://github.com/DestinE-Climate-DT/AQUA/compare/v0.19.0...v0.20.0
 [v0.19.0]: https://github.com/DestinE-Climate-DT/AQUA/compare/v0.18.1...v0.19.0